+++ CHANGELOG.md
@@ -10,6 +10,9 @@ ADDED:
 
 CHANGED:
 
+- update gradle to new version 7.3
+- update of bcpkix-jdk15on dependency version to 1.70
+- update of test dependency jobj-core to new version to 5
 - replaced null returns with Optional.empty()
 
 Version 7.9